refactor(test): move github helpers to db-test-seeders and db-test-assertions
Move all 14 GitHub test helper functions from api-test-helpers/github.ts
to their proper locations:
- 8 seeders to db-test-seeders/github.ts (with @why-db-direct annotations)
- 6 assertion helpers to db-test-assertions/github.ts

After this change, api-test-helpers/github.ts contains only re-exports
with zero db/schema imports.

Closes #9353
